ajeitei

+++ ex1-biblioteca-v4-SolExer/src/main/java/com/bcopstein/ex1biblioeca/Livro.java
@@ -1,19 +1,19 @@
 package com.bcopstein.ex1biblioeca;
 
 public class Livro {
-    private int id;
+    private long id;
     private String titulo;
     private String autor;
     private int ano;
 
-    public Livro(int id, String titulo, String autor, int ano) {
+    public Livro(long id, String titulo, String autor, int ano) {
         this.id = id;
         this.titulo = titulo;
         this.autor = autor;
         this.ano = ano;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
